// frontend/src/domain/clinician/insertClinician.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../api/graphql";

export type InsertClinicianVariables = {
  storeId: string;
  input: {
    id: string;
    code: string;
    initials: string;
    lastName: string;
    firstName?: string | null;
    gender?: "FEMALE" | "MALE" | "TRANSGENDER" | "TRANSGENDER_MALE" | "TRANSGENDER_MALE_HORMONE" | "TRANSGENDER_MALE_SURGICAL" | "TRANSGENDER_FEMALE" | "TRANSGENDER_FEMALE_HORMONE" | "TRANSGENDER_FEMALE_SURGICAL" | "UNKNOWN" | "NON_BINARY" | null;
    mobile?: string | null;
  };
};

export type InsertClinicianResult = {
  insertClinician: {
  id: string;
};
};

export const InsertClinician = {
  query: "mutation insertClinician($storeId: String!, $input: InsertClinicianInput!) {\n  insertClinician(storeId: $storeId, input: $input) {\n    id\n  }\n}",
} as TypedDocument<InsertClinicianResult, InsertClinicianVariables>;